shapemaker::geometry::region

Struct Region

pub struct Region {
    pub start: Point,
    pub end: Point,
}

Fields
start: Pointend: Point
Implementations

impl Region

pub fn iter(&self) -> RegionIterator

iterates from left to right then top to bottom (in a “row-major” order)

pub fn except<'a>( &self, except: &'a Region, ) -> impl Iterator<Item = Point> + use<'a>

Iterates all points except the ones specified in the except region

pub fn iter_lower_triangle(&self) -> impl Iterator<Item = Point>

pub fn iter_upper_strict_triangle(&self) -> impl Iterator<Item = Point>

pub fn outline(&self) -> impl Iterator<Item = Point>

Iterates all points outlining the region, in clockwise order starting from top-left

pub fn top_edge(&self) -> impl DoubleEndedIterator<Item = Point>

pub fn bottom_edge(&self) -> impl DoubleEndedIterator<Item = Point>

pub fn left_edge(&self) -> impl DoubleEndedIterator<Item = Point>

pub fn right_edge(&self) -> impl DoubleEndedIterator<Item = Point>

pub fn corners(&self) -> [Point; 4]

Corners of the region’s outline Does not match .bottomright() etc., since this method takes into account that the region is inclusive topleft, topright, bottomright, bottomleft

pub fn is_empty(&self) -> bool

pub fn ensure_nonempty(&self) -> Result<()>

pub fn rectangle(&self) -> Shape

pub fn center_coords(&self, cell_size: usize) -> (f32, f32)

impl Region

pub fn new( start: impl Into<Point>, end: impl Into<Point>, ) -> Result<Self, Error>

pub fn bottomleft(&self) -> Point

pub fn bottomright(&self) -> Point

pub fn topleft(&self) -> Point

pub fn topright(&self) -> Point

pub fn center(&self) -> Point

pub fn max<'a>(&'a self, other: &'a Region) -> &'a Region

pub fn merge<'a>(&'a self, other: &'a Region) -> Region

pub fn from_origin(end: Point) -> Result<Self>

pub fn from_topleft(origin: Point, size: (usize, usize)) -> Result<Self>

pub fn starting_from_topleft(&self, size: (usize, usize)) -> Result<Self>

pub fn from_bottomleft(origin: Point, (w, h): (usize, usize)) -> Result<Self>

pub fn starting_from_bottomleft(&self, size: (usize, usize)) -> Result<Self>

pub fn from_bottomright(origin: Point, (w, h): (usize, usize)) -> Result<Self>

pub fn starting_from_bottomright(&self, size: (usize, usize)) -> Result<Self>

pub fn from_topright(origin: Point, (w, h): (usize, usize)) -> Result<Self>

pub fn starting_from_topright(&self, size: (usize, usize)) -> Result<Self>

pub fn from_center_and_size( center: Point, (w, h): (usize, usize), ) -> Result<Self>

pub fn ensure_valid(self) -> Result<Self>

pub fn translate(&mut self, dx: i32, dy: i32)

pub fn translated(&self, dx: i32, dy: i32) -> Self

pub fn translated_by(&self, point: Point) -> Self

pub fn enlarged(&self, add_x: i32, add_y: i32) -> Self

adds dx and dy to the end of the region (dx and dy are not multiplicative but additive factors)

pub fn resized(&self, add_x: i32, add_y: i32) -> Self

resized is like enlarged, but transforms from the center, by first translating the region by (-dx, -dy)

pub fn x_range(&self) -> RangeInclusive<usize>

pub fn y_range(&self) -> RangeInclusive<usize>

pub fn x_range_without_last(&self) -> Range<usize>

pub fn y_range_without_last(&self) -> Range<usize>

pub fn within(&self, other: &Region) -> bool

pub fn clamped(&self, within: &Region) -> Region

pub fn width(&self) -> usize

pub fn height(&self) -> usize

pub fn dimensions(&self) -> (usize, usize)

pub fn size(&self, cell_size: usize) -> (f32, f32)

pub fn mirrored_width_range(&self) -> RangeInclusive<i32>

pub fn mirrored_height_range(&self) -> RangeInclusive<i32>

pub fn split(&self, along: Axis) -> (Region, Region)

impl Region

pub fn clip_path_id(&self) -> String

impl Region

pub fn random_end(&self, rng: &mut impl Rng, start: Point) -> Point

pub fn random(rng: &mut impl Rng, within: &Region) -> Self

pub fn random_point(&self, rng: &mut impl Rng) -> Point

pub fn random_point_except(&self, rng: &mut impl Rng, except: &Region) -> Point