shapemaker::graphics::canvas

Struct Canvas

pub struct Canvas {
    pub grid_size: (usize, usize),
    pub cell_size: usize,
    pub objects_count_range: Range<usize>,
    pub polygon_vertices_range: Range<usize>,
    pub outer_padding: usize,
    pub object_sizes: ObjectSizes,
    pub font_options: FontOptions,
    pub colormap: ColorMapping,
    pub name: String,
    pub layers: Vec<Layer>,
    pub background: Option<Color>,
    pub world_region: Region,
    /* private fields */
}

Fields
grid_size: (usize, usize)cell_size: usizeobjects_count_range: Range<usize>polygon_vertices_range: Range<usize>outer_padding: usizeobject_sizes: ObjectSizesfont_options: FontOptionscolormap: ColorMappingname: Stringlayers: Vec<Layer>

The layers are in order of top to bottom: the first layer will be rendered on top of the second, etc.

background: Option<Color>world_region: Region
Implementations

impl Canvas

pub fn new(width: usize, height: usize) -> Self

pub fn default_settings() -> Self

pub fn with_layers(layer_names: Vec<&str>) -> Self

Create a new canvas. The layers are in order of top to bottom: the first layer will be rendered on top of the second, etc. A layer named “root” will be added below all layers if you don’t add it yourself.

pub fn with_colors(colormap: ColorMapping) -> Self

pub fn init_layers(&mut self, names: Vec<&str>)

pub fn set_grid_size(&mut self, new_width: usize, new_height: usize)

pub fn layer(&mut self, name: &str) -> Result<&mut Layer>

pub fn layer_unchecked(&mut self, name: &str) -> &mut Layer

pub fn new_layer(&mut self, name: &str) -> &mut Layer

pub fn add_layer(&mut self, layer: Layer) -> &mut Layer

pub fn layer_or_empty(&mut self, name: &str) -> &mut Layer

pub fn layer_exists(&self, name: &str) -> bool

pub fn ensure_layer_exists(&self, name: &str)

pub fn put_layer_on_top(&mut self, name: &str)

puts this layer on top, and the others below, without changing their order

pub fn put_layer_on_bottom(&mut self, name: &str)

puts this layer on bottom, and the others above, without changing their order

pub fn reorder_layers(&mut self, new_order: Vec<&str>)

re-order layers. The first layer in the list will be on top, the last at the bottom

pub fn remove_layer(&mut self, name: &str)

pub fn root(&mut self) -> &mut Layer

pub fn add_object( &mut self, layer: &str, name: &str, shape: Shape, fill: Option<Fill>, ) -> Result<()>

pub fn remove_object(&mut self, name: &str)

pub fn set_background(&mut self, color: Color)

pub fn remove_background(&mut self)

pub fn fonts_loaded(&self) -> bool

pub fn load_fonts(&mut self) -> Result<()>

pub fn add_or_replace_layer(&mut self, layer: Layer)

pub fn region_is_whole_grid(&self, region: &Region) -> bool

pub fn clear(&mut self)

pub fn resolution_to_size_even(&self, resolution: u32) -> (u32, u32)

pub fn resolution_to_size(&self, resolution: u32) -> (u32, u32)

pub fn width(&self) -> usize

pub fn height(&self) -> usize

pub fn aspect_ratio(&self) -> f32

pub fn remove_all_objects_in(&mut self, region: &Region)

pub fn unique_filters(&self) -> Vec<Filter>

returns a list of all unique filters used throughout the canvas used to only generate one definition per filter

pub fn unique_pattern_fills(&self) -> Vec<Fill>

pub fn debug_region(&mut self, region: &Region, color: Color)

pub fn debug_grid(&mut self, color: Color)

pub fn dimensions<T: From<usize>>(&self) -> (T, T)

impl Canvas

pub fn random_layer(&mut self, rng: &mut impl Rng, name: &str) -> Layer

pub fn random_object(&mut self, rng: &mut impl Rng) -> Shape

pub fn random_object_within( &mut self, rng: &mut impl Rng, region: &Region, ) -> Shape

pub fn n_random_curves_within( &mut self, rng: &mut impl Rng, region: &Region, count: usize, layer_name: &str, ) -> Layer

pub fn random_curves_within( &mut self, rng: &mut impl Rng, layer_name: &str, region: &Region, object_counts: impl SampleRange<usize>, ) -> Layer

pub fn random_layer_within( &mut self, rng: &mut impl Rng, name: &str, region: &Region, ) -> Layer

pub fn random_linelikes( &mut self, rng: &mut impl Rng, layer_name: &str, ) -> Layer

pub fn random_region(&mut self, rng: &mut impl Rng) -> Region

impl Canvas

pub fn svg_to_pixmap( &self, width: u32, height: u32, contents: &str, ) -> Result<Pixmap>

pub fn render_to_pixmap(&mut self, width: u32, height: u32) -> Result<Pixmap>

pub fn render_to_png(&mut self, at: &str, resolution: u32) -> Result<()>

pub fn render_to_svg_string(&mut self) -> Result<String>

pub fn render_to_svg_file(&mut self, at: impl Into<PathBuf>) -> Result<()>

impl Canvas

pub fn show_available_fonts(&self)