shapemaker::rendering::renderable

Trait SVGAttributesRenderable

pub trait SVGAttributesRenderable {
    const MULTIPLE_VALUES_JOIN_BY: &'static str = ", ";

    // Required method
    fn render_to_svg_attributes(
        &self,
        colormap: ColorMapping,
        cell_size: usize,
        object_sizes: ObjectSizes,
        id: &str,
    ) -> Result<HashMap<String, String>>;
}

Struct can be rendered as attributes of a SVG element

Provided Associated Constants

const MULTIPLE_VALUES_JOIN_BY: &'static str = ", "

When merging multiple SVGAttributesRenderable, this string is used to join multiple values for the same key

Required Methods

fn render_to_svg_attributes( &self, colormap: ColorMapping, cell_size: usize, object_sizes: ObjectSizes, id: &str, ) -> Result<HashMap<String, String>>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.