shapemaker::video::easings

Enum StandardEasing

pub enum StandardEasing {

    InSine,
    OutSine,
    InOutSine,
    InQuadradic,
    OutQuadradic,
    InOutQuadradic,
    InCubic,
    OutCubic,
    InOutCubic,
    InQuartic,
    OutQuartic,
    InOutQuartic,
    InQuintic,
    OutQuintic,
    InOutQuintic,
    InExponential,
    OutExponential,
    InOutExponential,
    InCircular,
    OutCircular,
    InOutCircular,
    InBack,
    OutBack,
    InOutBack,
    InElastic,
    OutElastic,
    InOutElastic,
    InBounce,
    OutBounce,
    InOutBounce,
    Linear,

}

An enumeration of all strandard easings provided.

Variants

InSine

Eases in using a sine wave

See https://easings.net/#easeInSine for a visualization and more information.

OutSine

Eases out using a sine wave

See https://easings.net/#easeOutSine for a visualization and more information.

InOutSine

Eases in and out using a sine wave

See https://easings.net/#easeInOutSine for a visualization and more information.

InQuadradic

Eases in using a quadradic (x^2) curve

See https://easings.net/#easeInQuad for a visualization and more information.

OutQuadradic

Eases out using a quadradic (x^2) curve

See https://easings.net/#easeOutQuad for a visualization and more information.

InOutQuadradic

Eases in and out using a quadradic (x^2) curve

See https://easings.net/#easeInOutQuad for a visualization and more information.

InCubic

Eases in using a cubic (x^3) curve

See https://easings.net/#easeInCubic for a visualization and more information.

OutCubic

Eases out using a cubic (x^3) curve

See https://easings.net/#easeOutCubic for a visualization and more information.

InOutCubic

Eases in and out using a cubic (x^3) curve

See https://easings.net/#easeInOutCubic for a visualization and more information.

InQuartic

Eases in using a quartic (x^4) curve

See https://easings.net/#easeInQuart for a visualization and more information.

OutQuartic

Eases out using a quartic (x^4) curve

See https://easings.net/#easeOutQuart for a visualization and more information.

InOutQuartic

Eases in and out using a quartic (x^4) curve

See https://easings.net/#easeInOutQuart for a visualization and more information.

InQuintic

Eases in using a quintic (x^5) curve

See https://easings.net/#easeInQuint for a visualization and more information.

OutQuintic

Eases out using a quintic (x^5) curve

See https://easings.net/#easeOutQuint for a visualization and more information.

InOutQuintic

Eases in and out using a quintic (x^5) curve

See https://easings.net/#easeInOutQuint for a visualization and more information.

InExponential

Eases in using an expenential curve

See https://easings.net/#easeInExpo for a visualization and more information.

OutExponential

Eases out using an expenential curve

See https://easings.net/#easeOutExpo for a visualization and more information.

InOutExponential

Eases in and out using an expenential curve

See https://easings.net/#easeInOutExpo for a visualization and more information.

InCircular

Eases in using a curve resembling the top-left arc of a circle

See https://easings.net/#easeInCirc for a visualization and more information.

OutCircular

Eases out using a curve resembling the top-left arc of a circle

See https://easings.net/#easeOutCirc for a visualization and more information.

InOutCircular

Eases in and out using a curve resembling the top-left arc of a circle

See https://easings.net/#easeInOutCirc for a visualization and more information.

InBack

Eases in using a curve that backs away initially

See https://easings.net/#easeInBack for a visualization and more information.

OutBack

Eases out using a curve that backs away initially

See https://easings.net/#easeOutBack for a visualization and more information.

InOutBack

Eases in and out using a curve that backs away initially

See https://easings.net/#easeInOutBack for a visualization and more information.

InElastic

Eases in using a curve that bounces around the start initially then quickly accelerates

See https://easings.net/#easeInElastic for a visualization and more information.

OutElastic

Eases out using a curve that bounces around the start initially then quickly accelerates

See https://easings.net/#easeOutElastic for a visualization and more information.

InOutElastic

Eases in and out using a curve that bounces around the start initially then quickly accelerates

See https://easings.net/#easeInOutElastic for a visualization and more information.

InBounce

Eases in using a curve that bounces progressively closer as it progresses

See https://easings.net/#easeInBounce for a visualization and more information.

OutBounce

Eases out using a curve that bounces progressively closer as it progresses

See https://easings.net/#easeOutBounce for a visualization and more information.

InOutBounce

Eases in and out using a curve that bounces progressively closer as it progresses

See https://easings.net/#easeInOutBounce for a visualization and more information.

Linear

Eases linearly

Implementations

impl StandardEasing

pub fn all() -> &'static [StandardEasing]

Returns a collection of every variant of this enum.