shapemaker::video::engine

Struct Video

pub struct Video<C> {

    pub fps: usize,
    pub initial_canvas: Canvas,
    pub hooks: Vec<Hook<C>>,
    pub commands: Vec<Box<Command<C>>>,
    pub frames: Vec<Canvas>,
    pub frames_output_directory: &'static str,
    pub syncdata: SyncData,
    pub audiofile: PathBuf,
    pub resolution: u32,
    pub duration_override: Option<usize>,
    pub start_rendering_at: usize,
    pub progress_bar: ProgressBar,
    pub encoder: Option<Arc<Mutex<Encoder>>>,

}

Fields
fps: usizeinitial_canvas: Canvashooks: Vec<Hook<C>>commands: Vec<Box<Command<C>>>frames: Vec<Canvas>frames_output_directory: &'static strsyncdata: SyncDataaudiofile: PathBufresolution: u32duration_override: Option<usize>start_rendering_at: usizeprogress_bar: ProgressBarencoder: Option<Arc<Mutex<Encoder>>>
Implementations

impl<AdditionalContext: Default> Video<AdditionalContext>

pub fn new(canvas: Canvas) -> Self

pub fn sync_audio_with(self, sync_data_path: &str) -> Self

pub fn with_hook(self, hook: Hook<AdditionalContext>) -> Self

pub fn init( self, render_function: &'static RenderFunction<AdditionalContext>, ) -> Self

pub fn on( self, marker_text: &'static str, render_function: &'static RenderFunction<AdditionalContext>, ) -> Self

pub fn each_beat( self, render_function: &'static RenderFunction<AdditionalContext>, ) -> Self

pub fn every( self, amount: f32, unit: MusicalDurationUnit, render_function: &'static RenderFunction<AdditionalContext>, ) -> Self

pub fn each_frame( self, render_function: &'static RenderFunction<AdditionalContext>, ) -> Self

pub fn each_n_frame( self, n: usize, render_function: &'static RenderFunction<AdditionalContext>, ) -> Self

pub fn on_stem( self, stem_name: &'static str, threshold: f32, above_amplitude: &'static RenderFunction<AdditionalContext>, below_amplitude: &'static RenderFunction<AdditionalContext>, ) -> Self

threshold is a value between 0 and 1: current amplitude / max amplitude of stem

pub fn on_note( self, stems: &'static str, render_function: &'static RenderFunction<AdditionalContext>, ) -> Self

Triggers when a note starts on one of the stems in the comma-separated list of stem names stems.

pub fn on_note_end( self, stems: &'static str, render_function: &'static RenderFunction<AdditionalContext>, ) -> Self

Triggers when a note stops on one of the stems in the comma-separated list of stem names stems.

pub fn with_note( self, stems: &'static str, cutoff_amplitude: f32, layer_name: &'static str, object_name: &'static str, create_object: &'static dyn Fn(&Canvas, &mut Context<'_, AdditionalContext>) -> Result<ColoredObject>, ) -> Self

pub fn at_frame( self, frame: usize, render_function: &'static RenderFunction<AdditionalContext>, ) -> Self

pub fn when_remaining( self, seconds: usize, render_function: &'static RenderFunction<AdditionalContext>, ) -> Self

pub fn at_timestamp( self, timestamp: &'static str, render_function: &'static RenderFunction<AdditionalContext>, ) -> Self

pub fn command( self, command_name: &'static str, action: &'static CommandAction<AdditionalContext>, ) -> Self

pub fn bind_amplitude( self, layer: &'static str, stem: &'static str, update: &'static LayerAnimationUpdateFunction, ) -> Self

pub fn total_frames(&self) -> usize

pub fn duration_ms(&self) -> usize

pub fn setup_progress_bar(&self) -> ProgressBar

impl<AdditionalContext: Default> Video<AdditionalContext>

pub fn render_frames(&mut self) -> Result<usize>

pub fn render(&mut self, output_file: impl Into<PathBuf>) -> Result<()>