shapemaker::video::hooks

Trait AttachHooks

pub trait AttachHooks<C>: Sized {

    // Required method
    fn with_hook(self, hook: Hook<C>) -> Self;

    // Provided methods
    fn hook(
        self,
        when: &'static HookCondition<C>,
        render_function: &'static RenderFunction<C>,
    ) -> Self { ... }
    fn init(self, render_function: &'static RenderFunction<C>) -> Self { ... }
    fn dump_frame_when(self, when: &'static HookCondition<C>) -> Self { ... }
    fn on(
        self,
        marker_text: &'static str,
        render_function: &'static RenderFunction<C>,
    ) -> Self { ... }
    fn assign_scene_to(
        self,
        marker_text: &'static str,
        scene_name: &'static str,
    ) -> Self { ... }
    fn each_beat(self, render_function: &'static RenderFunction<C>) -> Self { ... }
    fn every(
        self,
        amount: f32,
        unit: MusicalDurationUnit,
        render_function: &'static RenderFunction<C>,
    ) -> Self { ... }
    fn each_frame(self, render_function: &'static RenderFunction<C>) -> Self { ... }
    fn each_n_frame(
        self,
        n: usize,
        render_function: &'static RenderFunction<C>,
    ) -> Self { ... }
    fn on_stem(
        self,
        stem_name: &'static str,
        threshold: f32,
        above_amplitude: &'static RenderFunction<C>,
        below_amplitude: &'static RenderFunction<C>,
    ) -> Self { ... }
    fn on_note(
        self,
        stems: &'static str,
        render_function: &'static RenderFunction<C>,
    ) -> Self { ... }
    fn on_note_end(
        self,
        stems: &'static str,
        render_function: &'static RenderFunction<C>,
    ) -> Self { ... }
    fn with_note<ObjectCreator>(
        self,
        stems: &'static str,
        cutoff_amplitude: f32,
        layer_name: &'static str,
        object_name: &'static str,
        create_object: &'static ObjectCreator,
    ) -> Self
       where ObjectCreator: Fn(&Canvas, &mut Context<'_, C>) -> Result<Object> + Send + Sync { ... }
    fn at_frame(
        self,
        frame: usize,
        render_function: &'static RenderFunction<C>,
    ) -> Self { ... }
    fn when_remaining(
        self,
        seconds: usize,
        render_function: &'static RenderFunction<C>,
    ) -> Self { ... }
    fn at_timestamp(
        self,
        timestamp: &'static str,
        render_function: &'static RenderFunction<C>,
    ) -> Self { ... }
    fn bind_amplitude(
        self,
        stem: &'static str,
        update: &'static AnimationUpdateFunction,
    ) -> Self { ... }

}

Required Methods

fn with_hook(self, hook: Hook<C>) -> Self

Provided Methods

fn hook( self, when: &'static HookCondition<C>, render_function: &'static RenderFunction<C>, ) -> Self

fn init(self, render_function: &'static RenderFunction<C>) -> Self

fn dump_frame_when(self, when: &'static HookCondition<C>) -> Self

fn on( self, marker_text: &'static str, render_function: &'static RenderFunction<C>, ) -> Self

fn assign_scene_to( self, marker_text: &'static str, scene_name: &'static str, ) -> Self

fn each_beat(self, render_function: &'static RenderFunction<C>) -> Self

fn every( self, amount: f32, unit: MusicalDurationUnit, render_function: &'static RenderFunction<C>, ) -> Self

fn each_frame(self, render_function: &'static RenderFunction<C>) -> Self

fn each_n_frame( self, n: usize, render_function: &'static RenderFunction<C>, ) -> Self

fn on_stem( self, stem_name: &'static str, threshold: f32, above_amplitude: &'static RenderFunction<C>, below_amplitude: &'static RenderFunction<C>, ) -> Self

threshold is a value between 0 and 1: current amplitude / max amplitude of stem

fn on_note( self, stems: &'static str, render_function: &'static RenderFunction<C>, ) -> Self

Triggers when a note starts on one of the stems in the comma-separated list of stem names stems.

fn on_note_end( self, stems: &'static str, render_function: &'static RenderFunction<C>, ) -> Self

Triggers when a note stops on one of the stems in the comma-separated list of stem names stems.

fn with_note<ObjectCreator>( self, stems: &'static str, cutoff_amplitude: f32, layer_name: &'static str, object_name: &'static str, create_object: &'static ObjectCreator, ) -> Self
where ObjectCreator: Fn(&Canvas, &mut Context<'_, C>) -> Result<Object> + Send + Sync,

fn at_frame( self, frame: usize, render_function: &'static RenderFunction<C>, ) -> Self

fn when_remaining( self, seconds: usize, render_function: &'static RenderFunction<C>, ) -> Self

fn at_timestamp( self, timestamp: &'static str, render_function: &'static RenderFunction<C>, ) -> Self

fn bind_amplitude( self, stem: &'static str, update: &'static AnimationUpdateFunction, ) -> Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.