shapemaker::video::video

Struct Video

pub struct Video<C> {

    pub fps: usize,
    pub initial_canvas: Canvas,
    pub hooks: Vec<Hook<C>>,
    pub commands: Vec<Box<Command<C>>>,
    pub frames: Vec<Canvas>,
    pub frames_output_directory: &'static str,
    pub syncdata: SyncData,
    pub audiofile: PathBuf,
    pub resolution: u32,
    pub duration_override: Option<Duration>,
    pub start_rendering_at: Timestamp,
    pub progress_bars: VideoProgressBars,
    pub progress: MultiProgress,

}

Fields
fps: usizeinitial_canvas: Canvashooks: Vec<Hook<C>>commands: Vec<Box<Command<C>>>frames: Vec<Canvas>frames_output_directory: &'static strsyncdata: SyncDataaudiofile: PathBufresolution: u32duration_override: Option<Duration>start_rendering_at: Timestampprogress_bars: VideoProgressBarsprogress: MultiProgress
Implementations

impl<C: Default> Video<C>

pub fn render( &self, output: SyncSender<EngineOutput>, controller: &EngineController<C>, ) -> Result<usize>

pub fn render_frame(&self, frame_no: usize, render_ahead: usize) -> Result<Node>

Render a single frame at the given frame number. Skip all hooks, expect for render_ahead frames before the requested one.

impl<C: Default> Video<C>

pub fn new(canvas: Canvas) -> Self

pub fn sync_audio_with(self, filepath: impl Into<PathBuf>) -> Result<Self>

pub fn ms_to_frames(&self, ms: usize) -> usize

pub fn duration_ms(&self) -> usize

pub fn constrained_ms_range(&self) -> Range<usize>

pub fn total_ms_range(&self) -> Range<usize>

pub fn total_duration_ms(&self) -> usize

Duration of the video, without taking into account a possible duration override.

pub fn with_scene(self, scene: Scene<C>) -> Self

Adds hooks from the given scene to the video. Hooks will be triggered when the current scene matches the scene’s name. Use Context#switch_scene to change scenes during rendering. See also with_marked_scene for a more ergonomic way to add scenes.

pub fn with_init_scene(self, scene: Scene<C>) -> Self

Adds the given scene and a hook that switches to it immediately.

pub fn with_marked_scene(self, scene: Scene<C>) -> Self

Adds the given scene, and a hook that switches to it when a marker with the same name is reached

pub fn command( self, command_name: &'static str, action: &'static CommandAction<C>, ) -> Self

impl<C: Default> Video<C>

pub fn setup_ffmpeg_encoder( &self, width: u32, height: u32, output_path: PathBuf, ) -> Result<FFMpegEncoder>

impl<C: Default> Video<C>

pub fn setup_vgv_encoder( &self, transcode: VGVTranscodeMode, width: usize, height: usize, initial_canvas: &Canvas, destination: impl Into<PathBuf>, ) -> Result<VGVEncoder>

impl<C: Default> Video<C>

pub fn encode( &mut self, output_file: impl Into<PathBuf> + Clone, ) -> Result<Duration>

pub fn encode_controlled( &mut self, output_file: impl Into<PathBuf> + Clone, engine_controller: &EngineController<C>, ) -> Result<Duration>

pub fn encode_with( &mut self, encoder: Box<dyn Encoder + Send>, engine_controller: &EngineController<C>, ) -> Result<Duration>

impl<C: 'static + Default> Video<C>

pub async fn serve(self, address: &str)